Moving ParseNote/FormatNote to utils, goblUnitFromUNECE to lines_parse.go
ParseNote/FormatNote have no UBL-specific logic -- they're a generic
#CODE#text note encoding, exported into utils alongside the other pure
helpers. goblUnitFromUNECE maps UN/ECE codes to GOBL units, used only
by lines_parse.go's quantity parsing, so it moves there instead of
utils -- it's UBL-adjacent unit-mapping logic, not a generic helper.

utils.go is now empty and removed.
