+++ utils/notes.go
@@ -1,4 +1,4 @@
-package ubl
+package utils
 
 import (
 	"fmt"
@@ -8,28 +8,16 @@ import (
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
-
-	"github.com/invopop/gobl.ubl/utils"
 )
-
-// goblUnitFromUNECE maps UN/ECE code to GOBL equivalent.
-func goblUnitFromUNECE(unece cbc.Code) org.Unit {
-	for _, def := range org.UnitDefinitions {
-		if def.UNECE == unece {
-			return def.Unit
-		}
-	}
-	return org.Unit(unece)
-}
 
 // noteCodePattern matches the #CODE#text format used in UBL notes to encode
 // UNTDID 4451 text subject qualifier codes, e.g. "#AAI#some text".
 var noteCodePattern = regexp.MustCompile(`^#([A-Z0-9]+)#(.*)$`)
 
-// parseNote converts a raw UBL note string into a GOBL Note. If the string
+// ParseNote converts a raw UBL note string into a GOBL Note. If the string
 // matches the #CODE#text format the code is stored as the untdid text-subject ext.
-func parseNote(text string) *org.Note {
-	text = utils.CleanString(text)
+func ParseNote(text string) *org.Note {
+	text = CleanString(text)
 	if m := noteCodePattern.FindStringSubmatch(text); m != nil {
 		return &org.Note{
 			Ext:  tax.ExtensionsOf(cbc.CodeMap{untdid.ExtKeyTextSubject: cbc.Code(m[1])}),
@@ -39,7 +27,8 @@ func parseNote(text string) *org.Note {
 	return &org.Note{Text: text}
 }
 
-func formatNote(note *org.Note) string {
+// FormatNote renders a GOBL note as its UBL note text.
+func FormatNote(note *org.Note) string {
 	if note == nil {
 		return ""
 	}
